WS2/requirements.txt
# ws1/requirements.txt e ws2/requirements.txt
Flask>=2.0
spyne>=2.13
lxml>=4.6
mysql-connector-python>=8.0
argon2-cffi>=21.1
requests>=2.25
gunicorn>=20.1
python-dotenv>=0.19
Werkzeug>=2.0 